yolo/scripts/requirements-ml-ws.txt
websocket-client>=1.6.0
python-dotenv>=1.0.0